Fixed random being int erroring later.

+++ CYOA Engine/src/nbradham/cyoa/CYOAEngine.java
@@ -283,7 +283,7 @@ final class CYOAEngine {
                     });
                 }
                 case ">r" -> {
-                    vars.put(nextTok(), r.nextInt(Integer.valueOf(nextTok()), Integer.valueOf(nextTok())));
+                    vars.put(nextTok(), (float) r.nextInt(Integer.valueOf(nextTok()), Integer.valueOf(nextTok())));
                 }
                 case ">t" -> {
                     testExp(() -> {
